+++ crates/sui-graphql-rpc/src/types/balance.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use super::{big_int::BigInt, move_type::MoveType};
-use crate::types::owner::Owner;
 use async_graphql::*;
 
 #[derive(Clone, Debug, SimpleObject)]
@@ -13,11 +12,4 @@ pub(crate) struct Balance {
     pub(crate) coin_object_count: Option<u64>,
     /// Total balance across all coin objects of the coin type
     pub(crate) total_balance: Option<BigInt>,
-}
-
-#[derive(Clone, Debug, SimpleObject)]
-pub(crate) struct BalanceChange {
-    pub(crate) owner: Option<Owner>,
-    pub(crate) amount: Option<BigInt>,
-    pub(crate) coin_type: Option<MoveType>,
 }
