+++ crates/sui-graphql-rpc/src/types/mod.rs
@@ -27,7 +27,6 @@ pub(crate) mod move_type;
 pub(crate) mod move_value;
 pub(crate) mod name_service;
 pub(crate) mod object;
-pub(crate) mod object_change;
 pub(crate) mod open_move_type;
 pub(crate) mod owner;
 pub(crate) mod protocol_config;
@@ -40,6 +39,7 @@ pub(crate) mod sui_address;
 pub(crate) mod sui_system_state_summary;
 pub(crate) mod system_parameters;
 pub(crate) mod transaction_block;
+pub(crate) mod transaction_block_effects;
 pub(crate) mod transaction_block_kind;
 pub(crate) mod transaction_signature;
 pub(crate) mod validator;
